+++ src/SmeehanBlogApi/Startup.cs
@@ -3,6 +3,7 @@ using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
+using SmeehanBlogApi.DynamoDB;
 using SmeehanBlogApi.Progress;
 using SmeehanBlogApi.Quotes;
 
@@ -24,9 +25,9 @@ namespace SmeehanBlogApi
         {
             services.AddControllers();
 
+            services.AddDynamoDBConnection(_config, _env);
             services.AddQuotes(_config, _env);
-
-            services.AddScoped<IProjectStore, ProjectStore>();
+            services.AddProgress(_config, _env);
 
             services.AddCors(o => o.AddPolicy("MyPolicy", localBuilder =>
             {